requirements.txt
# Dipendenze dell'applicazione
# FastAPI e dipendenze
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticazione
python-jose==3.3.0
cryptography==42.0.5
passlib==1.7.4
python-multipart==0.0.9

# Integrazione OpenAI
openai==1.12.0
langchain==0.1.10
langchain-openai==0.0.5

# Database vettoriale (opzionale)
qdrant-client==1.7.3

# Sintesi vocale (opzionale)
elevenlabs==0.2.27

# Utilità
python-dotenv==1.0.1
httpx==0.27.0